requirements.txt
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
babel==2.16.0
beautifulsoup4==4.12.3
bibtexparser==2.0.0b8
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clldutils==3.24.0
colorama==0.4.6
colorlog==6.9.0
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.0
csvw==3.5.1
cycler==0.12.1
daphne==4.1.2
deprecation==2.1.0
Django==5.1
django-bootstrap5==24.2
django-countries==7.6.1
django-environ==0.11.2
dlinfo==2.0.0
docx==0.2.4
duckduckgo_search==7.2.1
et_xmlfile==2.0.0
filelock==3.16.1
fonttools==4.55.8
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.27.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
isodate==0.7.2
Jinja2==3.1.5
joblib==1.4.2
json_repair==0.37.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kombu==5.4.2
lancedb==0.18.0
language-tags==1.2.0
llvmlite==0.44.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
munch==4.0.0
narwhals==1.24.2
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==2.1.3
ollama==0.4.6
openai-whisper==20240930
opencv-python==4.11.0.86
openpyxl==3.1.5
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.5
phonemizer==3.3.0
pillow==10.4.0
pip-autoremove @ git+https://github.com/imba-tjd/pip-autoremove@aaaf6240a09ed5820cb26b1e7507d55944a9cc75
primp==0.11.0
prompt_toolkit==3.0.50
protobuf==5.29.3
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pylance==0.22.0
pylatexenc==2.10
PyMuPDF==1.25.2
pyOpenSSL==25.0.0
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
rdflib==7.1.3
redis==5.2.1
referencing==0.36.1
regex==2024.11.6
reportlab==4.2.2
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
segments==2.2.1
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.0
soupsieve==2.6
sqlparse==0.5.1
streamlit==1.41.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
zope.interface==7.2
